split mystd & rpifunc
