+++ lab3/include/mystd.h
@@ -1,7 +1,3 @@
 int strcmp(char *s1, char *s2);
-double get_time();
-void reset(int tick);
-void cancel_reset();
 int atoi(char *str);
-void print_next_instr_addr();
-void show_boot_msg();
+
