+++ lab3/include/rpifunc.h
@@ -1,0 +1,4 @@
+double get_time();
+void reset(int tick);
+void cancel_reset();
+void show_boot_msg();
